finish follow

+++ apps/example/model/follow_model.go
@@ -9,7 +9,7 @@ type (
 		ID           	 int32 `gorm:"primary_key"`
 		UserId     		 int32 `json:"user_id"`
 		FollowBy     	 int32 `json:"followby"`
-		Ctime        	 int32  `json:"ctime"`
+		Ctime        	 int64  `json:"ctime"`
 	}
 )
